refactor: login button to new file
